feat: add database and update logic

+++ api-service/src/main/java/hibernate/v2/api/model/openweather/Minutely.kt
@@ -10,5 +10,5 @@ data class Minutely(
     @SerializedName("dt")
     val dt: Int? = null,
     @SerializedName("precipitation")
-    val precipitation: Int? = null
+    val precipitation: Double? = null
 ) : Parcelable
